hhvm compatability fix, add hhvm to travis #10

+++ Config/Schema/tasks.sql
@@ -2,8 +2,8 @@ delimiter $$
 
 CREATE TABLE `tasks` (
   `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
-  `process_id` int(10) DEFAULT 0,
-  `server_id` int(10) DEFAULT 0,
+  `process_id` bigint(20) DEFAULT 0,
+  `server_id` bigint(20) DEFAULT 0,
   `command` varchar(500) DEFAULT NULL,
   `path` varchar(500) DEFAULT NULL,
   `arguments` longtext,
